Backend/App/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.1.2
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
